REFACTOR: move web socket to own module
