+++ tests/whist/server/web_socket/test_side_channel.py
@@ -1,9 +1,9 @@
 from unittest import TestCase
 from unittest.mock import MagicMock
 
-from whist.server.api.publish.events.event import Event
-from whist.server.api.publish.side_channel import SideChannel
-from whist.server.api.publish.subscriber import Subscriber
+from whist.server.web_socket.events.event import Event
+from whist.server.web_socket.side_channel import SideChannel
+from whist.server.web_socket.subscriber import Subscriber
 
 
 class TestCase(TestCase):
