+++ whist/server/web_socket/side_channel.py
@@ -1,6 +1,6 @@
 """Handles push of events"""
-from whist.server.api.publish.events.event import Event
-from whist.server.api.publish.subscriber import Subscriber
+from whist.server.web_socket.events.event import Event
+from whist.server.web_socket.subscriber import Subscriber
 
 
 class SideChannel:
